rest

+++ src/main/java/niu/itmo/spaceresearch/repository/ResearcherRepository.java
@@ -2,7 +2,6 @@ package niu.itmo.spaceresearch.repository;
 
 import niu.itmo.spaceresearch.model.Researcher;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 
 import java.util.Optional;
 
@@ -10,6 +9,7 @@ import java.util.Optional;
  * @author amifideles
  */
 
-public interface ResearcherRepository extends JpaRepository<Researcher, Long> {
+public interface ResearcherRepository extends JpaRepository<Researcher, Integer> {
     Optional<Researcher> findByUsername(String username);
+
 }
